requirements.txt
fastapi[all]==0.89.1
uvicorn[standard]==0.20.0
sqlalchemy==1.4.46
pydantic==1.10.4
psycopg2-binary==2.9.5
bcrypt==4.0.1
alembic==1.9.1
passlib==1.7.4
python-dotenv==0.21.0
python-jose==3.3.0
boto3==1.26.49
fastapi-pagination==0.11.2
openpyxl==3.1.2